+++ examples/fishy2.rs
@@ -1,12 +1,12 @@
 // fishy2.rs
-use footile::{FillRule, PathBuilder, Plotter};
+use footile::{FillRule, Path, Plotter};
 use pix::rgb::{Rgba8p, SRgba8};
 use pix::Raster;
 
 mod png;
 
 fn main() -> Result<(), std::io::Error> {
-    let fish = PathBuilder::default()
+    let fish = Path::default()
         .relative()
         .pen_width(3.0)
         .move_to(112.0, 24.0)
@@ -15,15 +15,15 @@ fn main() -> Result<(), std::io::Error> {
         .line_to(32.0, 24.0)
         .line_to(-16.0, -40.0)
         .close()
-        .build();
-    let eye = PathBuilder::default()
+        .finish();
+    let eye = Path::default()
         .relative()
         .pen_width(2.0)
         .move_to(24.0, 48.0)
         .line_to(8.0, 8.0)
         .move_to(0.0, -8.0)
         .line_to(-8.0, 8.0)
-        .build();
+        .finish();
     let v = vec![Rgba8p::new(0, 0, 0, 0); 128 * 128];
     let r = Raster::<Rgba8p>::with_pixels(128, 128, v);
     let mut p = Plotter::new(r);