+++ examples/teeth.rs
@@ -1,11 +1,11 @@
-use footile::{PathBuilder, Plotter};
+use footile::{Path, Plotter};
 use pix::matte::Matte8;
 use pix::Raster;
 
 mod png;
 
 fn main() -> Result<(), std::io::Error> {
-    let path = PathBuilder::default()
+    let path = Path::default()
         .relative()
         .move_to(0.0, 8.0)
         .line_to(8.0, 8.0)
@@ -25,7 +25,7 @@ fn main() -> Result<(), std::io::Error> {
         .line_to(8.0, -8.0)
         .line_to(8.0, 8.0)
         .line_to(8.0, -8.0)
-        .build();
+        .finish();
     let mut p = Plotter::new(Raster::with_clear(64, 64));
     png::write_matte(p.stroke(&path, Matte8::new(255)), "./teeth.png")
 }